refactor: move executor to helper and reuse it
