+++ tests/helper.py
@@ -1,3 +1,16 @@
+from docarray import DocumentArray
+
+from jina import Executor, requests
+
+
+class ProcessExecutor(Executor):
+    @requests(on='/')
+    def process(self, docs: DocumentArray, **kwargs):
+        for doc in docs:
+            doc.text = doc.text + 'world'
+            doc.tags['processed'] = True
+
+
 def _validate_dummy_custom_gateway_response(port, expected):
     import requests
